requirements.txt
streamlit
PyMuPDF
langchain
langchain-community
sentence-transformers
faiss-cpu
python-dotenv
torch
asyncio